mid1_p2: Changed the max and min initial variables, set userNum greater than or equal to and less than or equal to the respective values.
Adjusted the max and min values by initializing with 0 and -1 respectively. When ran through the while loop, will update accordingly. Could not tell you why I chose 0 and -1, they just seem to work for the conditions. Actually, I don't think I even need a min variable.

Set the userNum >= max so if 0 is the first number inputted and the second number is negative like -10, the function will still output 0.

Signed-off-by: CieranW <cieranwong@gmail.com>