mid1_p3: Moved the printf statement and added another if statement in the first one under the while loop
Moved the first printf statement to the start of the while loop so that the first number (user input) is always printed. Added a second if statement inside the first so that when userNum == 1, skips the while loop and goes immediately to the second printf statement, printing a new line as well.

Signed-off-by: CieranW <cieranwong@gmail.com>