lab6: Init Commit
Init commit for Lab 6.

Problem 1: Given a file of integers, find the sum of all integers greater than or equal to an integer inputted by the user.

Problem 2: Write a program that can find the amount of money (in dollars and cents) from a given number of coins (quarter, dime, nickel, and penny) entered by the user. Must use the following function:
	void Credit(int quarter, int dime, int nickel, int penny, int *dollars, int *cents);

Signed-off-by: CieranW <cieranwong@gmail.com>