Midterm1: Init commit
Initial commit for Midterm 1 Part 1. Program to find instances of both upper and lower case letters in a given .txt file. Requirement is to print out the number of occurences in each line. If the userinput is lowercase t, the program must also count occurences of uppercase t (T) as well.

Signed-off-by: CieranW <cieranwong@gmail.com>