HW6P1: Added substance to the DistConverter function and main
In DistConverter, took a single input and changed it into yards, feet, and inches using some simple math. The yard took the amount of feet and divided by 3 and then feet was reassigned to be the remainder of yards.

In main, declared a file and int variables, called the DistConverter function and printed out the output from DistConverter.

Signed-off-by: CieranW <cieranwong@gmail.com>