HW6P2: Filled out the smallestInt function and main
In main, declared the file, called the function with pointers, and then priinted out the result.

In smallestInt, took the input, scanned into a temporary variable and in the while loop, used if else statements to adjust based on whether or not the number scanned was the smaller or second smallest value.

Signed-off-by: CieranW <cieranwong@gmail.com>