HW7 Part 1: Init Commit
Initial commit for HW 7 Part 1. Purpose of the program is explained below:

Assume that there are 1,000 bins numbered from 1 to 1000 and 100 players numbered from 1 to 100. Initially, all bins are empty. A player with id x visits all bins that are a multiple of her id. The player will throw a ball in a bin that she visits if the bin is empty, or pick up a ball if the bin has at least one ball. After all 100 players complete their turn, we want to find the number and ids of the empty bins. Write a C program that mimics this experiment and prints a) the number of empty bins, and b) the ids of the empty bins.

Signed-off-by: CieranW <cieranwong@gmail.com>