HW7 Part 1: Filled out the main function with four sets of for loops.
The first loop set each index in the bins array to 0, to signify that they were empty.

The second for loop was a nested for loop with an if else statement. The first for loop represented the players and each one went through the entire bins array looking for multiples of their number (second for loop). The if else statement was to identify if the bins were empty or not and to add/remove items based on the condition met.

Third for loop was to find the position of each bin that was empty and to start counting how many bins are empty.

The last for loop is to print out each position.

Signed-off-by: CieranW <cieranwong@gmail.com>