HW7 Part2: Added a new function to compare arrays and filled out the two og functions.
In teh bool function, letterCheck, took the original array holding the word you would like someone to guess, an array that held the guessed letters in the correct index, the length of the word, user inputted guess, and the number of times the letter appears in the word. There was an issue with times being unused as a pointer so I changed the incrementation of *times++; into *times += 1; and that seemed to solve the problem.

The compareArray function I found on Stack. It does exactly what it sounds like. I used this function to compare the original array with the guess array. When both arrays match, the while loop in main ends.

In main, the while loop was the meat of the function as it kept going till there were no tries left or the word was guessed. As mentioned above, the compareArray was utilised to see if the arrays matched. At the end of the function, the output varies based on the condition you met to exit the while loop. No more tries gave you a message saying you failed, and solving the word congratulated you.

For my personal note: Took me a while to figure this one out, but I do like the use of boolean and various other functions to ease the burden of having to continuously write a comparison statement through for loops. The pointer thing was finicky so might need to take another look at it.

Signed-off-by: CieranW <cieranwong@gmail.com>