Lab8 Part1: Finished writing ReadArray
ReadArray takes in a file input, scans to see if the numbers in a square shape add up to a magic square and if so, calculates the sum of numbers in a column, row and diagonal. Found the sum of numbers in a row and column by doing a nested for loop and using an if statement to see if they match. Simply reversed the variables used for the comparison. Ex: x[i][j] == x[j][i].

Also added the file needed for the program

Signed-off-by: CieranW <cieranwong@gmail.com>