Hw8: Wrote the first part of the boolean function
Began writing the boolean function to check if the sudoku puzzle from the input file is a valid puzzle. Started by assigning three different arrays a row, column, or square of the original arrray x. Then will most likely use if else statements to check if the values in each row, column, and square array repeat. Still not 100% sure on how to do this just yet.

Signed-off-by: CieranW <cieranwong@gmail.com>