HW 8: Finished the sudoku program
Finished writing this damn thing. Took a lot logner than expected and a lil too many moving parts so it threw me off a little bit. Figured out the square 3x3 comparison section. Needed an additonal function to store the values in to use for comparison later on. Turned out the thing that was throwing errors left and right was my temporary assignment to various variables. Take temprow for example. When assigning it, I mistakenly did temprow[row[i - 1]] = 1; which was incorrect and should have been temprow[row[i] - 1] = 1;. So confusing but I'm glad it's done with. May try some testcases in the future to see if it works for inaccurate files. Also check sudoku1.txt for the file I used to test.

I really do dislike C. It's getting annoying at this point and between you and me, my instructor isn't helping a ton. I know he's trying to get me to figure this out on my own, but even the hints aren't that great. Like the example above that I did before, that was his idea and I had to convert it in order for it to work. Time for bed.

Signed-off-by: CieranW <cieranwong@gmail.com>