Exam 2 Part 1: Init commit
Initial commit for Exam 2 part 1. Program to check for symmetry in a file. Assigns the data from the file into an array and runs through two functions. First to populate a designated array and second to check if it's symmetric.

If it's symmetric, outputs the pivot point in the array.

Signed-off-by: CieranW <cieranwong@gmail.com>