Final Project: Init Commit
Final Project for ECE175. Goal is to write a program for a DOS game. Similar to UNO but not quite the same. Currently in the folder is the main file, a header with the typedef struct, a C file for the add and remove functions, and a .txt file with the needed cards.

Signed-off-by: CieranW <cieranwong@gmail.com>